# python-sdk/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@katanemo.com
    name: Katanemo Labs Inc.
    url: https://api.katanemo.com
  description: "With Katanemo developers can add support for users, enterprise SSO,\
    \ machine keys and fine-grained authorization in minutes."
  termsOfService: https://www.katanemo.com/terms-of-service
  title: "Katanemo - identity, and fine-grained authorization for modern { API-first\
    \ } software companies."
  version: 1.0.0
servers:
- url: https://api.katanemo.com
...
paths:
...
  /org:
    get:
      description: Returns an object with information regarding an account which is
        present in the token
      operationId: GetAccountOrganization
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
          description: Organization of the acccount.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List Organizations
      tags:
      - organization
  /org/{accountId}:
    get:
      description: Returns an object with information regarding an account
      operationId: GetAccountInfo
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
          description: Organization
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get Organization
      tags:
      - organization
    put:
      description: Returns status code for successful or failed update.
      operationId: UpdateAccountInfo
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationRequest'
        description: Update account object
        required: true
      responses:
        "200":
          description: Successfully updated organization
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Update Organization
      tags:
      - organization
  /org/{accountId}/verify:
    get:
      description: "Triggers the domain verification flow. If TXT record is created\
        \ and has the correct verification code, the domain is verified."
      operationId: VerifyAccountDomain
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Domain verified
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Verify Domain
      tags:
      - organization
  /org/{accountId}/user:
    get:
      description: Returns a list of all users belonging to provided organization
        ID
      operationId: GetUsersForAccount
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserResponse'
                type: array
          description: users belonging to provided account ID
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List Users
      tags:
      - organization
      - identity
    post:
      description: "Creates a new User and triggers an email verification workflow,\
        \ followed by set-password"
      operationId: createUserForAccount
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        description: User information
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: User
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Invite User
      tags:
      - organization
      - identity
  /org/{accountId}/user/{userId}:
    get:
      description: Get a specific user for a particular organization
      operationId: GetUser
      parameters:
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: user belonging to provided account ID
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get User
      tags:
      - organization
      - identity
    put:
      description: Updates a User account
      operationId: updateUser
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: accountId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        description: User information
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: User
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Update user
      tags:
      - organization
      - identity
...
components:
  schemas:
...
    OrganizationResponse:
      example:
        samlConnectionsCount: 5
        rolesCount: 6
        oidcConnectionsCount: 1
        launchedServices:
        - launchedServices
        - launchedServices
        subscribedServices:
        - subscribedServices
        - subscribedServices
        subscribers:
        - subscribers
        - subscribers
        domainVerificationCode: domainVerificationCode
        usersCount: 0
        defaultConnectionType: defaultConnectionType
        accountId: accountId
        domainVerified: true
        adminAccount: adminAccount
        domain: domain
        defaultConnection: defaultConnection
        name: name
      properties:
        accountId:
          title: accountId
          type: string
        name:
          title: name
          type: string
        adminAccount:
          title: adminAccount
          type: string
        domainVerificationCode:
          title: domainVerificationCode
          type: string
        domain:
          title: domain
          type: string
        domainVerified:
          title: domainVerified
          type: boolean
        usersCount:
          title: usersCount
          type: integer
        rolesCount:
          title: rolesCount
          type: integer
        oidcConnectionsCount:
          title: oidcConnectionsCount
          type: integer
        samlConnectionsCount:
          title: samlConnectionsCount
          type: integer
        defaultConnection:
          title: defaultConnection
          type: string
        defaultConnectionType:
          title: defaultConnectionType
          type: string
        launchedServices:
          items:
            type: string
          title: launchedServices
          type: array
        subscribedServices:
          items:
            type: string
          title: subscribedServices
          type: array
        subscribers:
          items:
            type: string
          title: subscribers
          type: array
      required:
      - accountId
      - adminAccount
      - name
      title: OrganizationResponse
...
    Error:
      example:
        code: 0
        message: message
      properties:
        code:
          description: Error code
          format: int32
          title: code
          type: integer
        message:
          description: Error message
          title: message
          type: string
      required:
      - code
      - message
      title: Error
    BadRequestException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: BadRequestException
      type: object
    UnauthorizedException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: UnauthorizedException
      type: object
...
    ConflictException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: ConflictException
      type: object
    TooManyRequestsException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: TooManyRequestsException
      type: object
    InternalServerErrorException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: InternalServerErrorException
      type: object
...